+++ setup.cfg
@@ -4,4 +4,4 @@ exclude =
     .git,
     __pycache__
 max-complexity = 10
-max-line-length = 120
+max-line-length = 150
